GH-1113 green-button region-connector: Move MeterReading to a dedicated package to prevent confusion with events
