+++ region-connectors/region-connector-us-green-button/src/main/java/energy/eddie/regionconnector/us/green/button/permission/handlers/AcceptedHandler.java
@@ -3,9 +3,9 @@ package energy.eddie.regionconnector.us.green.button.permission.handlers;
 import energy.eddie.regionconnector.shared.event.sourcing.EventBus;
 import energy.eddie.regionconnector.shared.event.sourcing.handlers.EventHandler;
 import energy.eddie.regionconnector.us.green.button.config.GreenButtonConfiguration;
-import energy.eddie.regionconnector.us.green.button.permission.events.MeterReading;
 import energy.eddie.regionconnector.us.green.button.permission.events.UsAcceptedEvent;
 import energy.eddie.regionconnector.us.green.button.permission.request.api.UsGreenButtonPermissionRequest;
+import energy.eddie.regionconnector.us.green.button.permission.request.meter.reading.MeterReading;
 import energy.eddie.regionconnector.us.green.button.persistence.UsPermissionRequestRepository;
 import energy.eddie.regionconnector.us.green.button.services.PermissionRequestService;
 import energy.eddie.regionconnector.us.green.button.services.historical.collection.HistoricalCollectionService;
