+++ region-connectors/region-connector-us-green-button/src/main/java/energy/eddie/regionconnector/us/green/button/services/historical/collection/HistoricalCollectionService.java
@@ -4,10 +4,10 @@ import energy.eddie.regionconnector.us.green.button.api.GreenButtonApi;
 import energy.eddie.regionconnector.us.green.button.api.Pages;
 import energy.eddie.regionconnector.us.green.button.client.dtos.meter.HistoricalCollectionResponse;
 import energy.eddie.regionconnector.us.green.button.client.dtos.meter.Meter;
-import energy.eddie.regionconnector.us.green.button.permission.events.MeterReading;
-import energy.eddie.regionconnector.us.green.button.permission.events.MeterReadingPk;
 import energy.eddie.regionconnector.us.green.button.permission.events.PollingStatus;
 import energy.eddie.regionconnector.us.green.button.permission.request.api.UsGreenButtonPermissionRequest;
+import energy.eddie.regionconnector.us.green.button.permission.request.meter.reading.MeterReading;
+import energy.eddie.regionconnector.us.green.button.permission.request.meter.reading.MeterReadingPk;
 import energy.eddie.regionconnector.us.green.button.persistence.MeterReadingRepository;
 import jakarta.annotation.Nullable;
 import org.slf4j.Logger;
